==> 1b. get_data_via_yadisk.ipynb <==
# Получение данных тестового счетчика с `Яндекс.Диска`
Для показательного очного семинара/воркшопа, или, если не получается по каким-то причинам получить данные из `LogsAPI`, были заготовлены CSV с выгруженными данными из `LogsAPI` тестового счетчика.

Хиты: https://disk.yandex.ru/d/MoSoXW6VzAP0bQ
Визиты: https://disk.yandex.ru/d/ywvr6YM3B95i0A

Нужно эти файлы скачать и положить в папочку репозитория для дальнейшей работы.

В файле `some_funcs` есть метод `get_file_from_yadisk(file_link, file_name)`, который копирует ровно 1 файл по ссылке из Яндекс.Диска `file_link` и сохраняет его на диск по пути `file_name`

Ожидается, что, выполняя ячейки ниже, вы будете находится в директории с репозиторием (`/yandex_metrika_cloud_case`)


import some_funcs
from some_funcs import get_file_from_yadisk

get_file_from_yadisk(file_link = 'https://disk.yandex.ru/d/MoSoXW6VzAP0bQ',
                    file_name = 'metrika_cloud_case_data_hits.csv')

get_file_from_yadisk(file_link = 'https://disk.yandex.ru/d/ywvr6YM3B95i0A',
                    file_name = 'metrika_cloud_case_data_visits.csv')
